requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-crispy-forms==1.14.0
django-rosetta==0.9.8
django-tinymce==3.4.0
idna==3.3
Pillow==9.0.1
polib==1.1.1
python-gettext==4.0
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
